requirements.txt
Flask
flask-cors
requests
google-generativeai
python-dotenv
torch
numpy
Pillow
scikit-learn
torchvision
transformers
